event-transformer: extract grid controller
