+++ src/apps/selection/app.rs
@@ -136,11 +136,9 @@ mod test {
 
     struct Transformer {}
     impl EventTransformer for Transformer {
-        fn get_grid_size(&self) -> Result<(usize, usize), Error> { Err(Error::Unsupported) }
         fn into_index(&self, _event: Event) -> Result<Option<u16>, Error> { Err(Error::Unsupported) }
         fn into_app_index(&self, _event: Event) -> Result<Option<u16>, Error> { Err(Error::Unsupported) }
         fn into_color_palette_index(&self, _event: Event) -> Result<Option<u16>, Error> { Err(Error::Unsupported) }
-        fn into_coordinates(&self, _event: Event) -> Result<Option<(u16, u16)>, Error> { Err(Error::Unsupported) }
         fn from_image(&self, _image: Image) -> Result<Event, Error> { Err(Error::Unsupported) }
         fn from_index_to_highlight(&self, _index: u16) -> Result<Event, Error> { Err(Error::Unsupported) }
         fn from_app_colors(&self, app_colors: Vec<[u8; 3]>) -> Result<Event, Error> {
