+++ src/main.rs
@@ -1,3 +1,5 @@
+#![feature(specialization)]
+
 #[macro_use] extern crate async_trait;
 #[macro_use] extern crate mockall;
