+++ src/midi/device.rs
@@ -5,6 +5,7 @@ use portmidi::{InputPort, OutputPort, MidiEvent, MidiMessage};
 
 pub use crate::image::Image;
 use super::Error;
+use super::features::GridController;
 
 #[derive(Clone, Debug, PartialEq)]
 pub enum Event {
@@ -12,11 +13,7 @@ pub enum Event {
     SysEx(Vec<u8>),
 }
 
-pub trait EventTransformer {
-    /// Device that can expose the size of its orthogonal grid,
-    /// as it should typically be the case for pads controllers.
-    fn get_grid_size(&self) -> Result<(usize, usize), Error>;
-
+pub trait EventTransformer: GridController {
     /// Device that can associate a MIDI event to an unsigned integer,
     /// that can be used to access elements of an indexed collections.
     fn into_index(&self, event: Event) -> Result<Option<u16>, Error>;
@@ -28,10 +25,6 @@ pub trait EventTransformer {
     /// Device that can associate a MIDI event to a color from a palette,
     /// let’s be realistic, this only be used by the paint application.
     fn into_color_palette_index(&self, event: Event) ->  Result<Option<u16>, Error>;
-
-    /// Device that can associate a MIDI event to a pair of coordinates,
-    /// this method can be used to select the pixel of an image; (0, 0) being the top-left corner.
-    fn into_coordinates(&self, event: Event) -> Result<Option<(u16, u16)>, Error>;
 
     /// Device that can render an image.
     fn from_image(&self, image: Image) -> Result<Event, Error>;
