+++ src/midi/devices/default/mod.rs
@@ -7,10 +7,6 @@ pub fn transformer() -> &'static DefaultEventTransformer {
 const DEFAULT_EVENT_TRANSFORMER: DefaultEventTransformer = DefaultEventTransformer {};
 pub struct DefaultEventTransformer {}
 impl EventTransformer for DefaultEventTransformer {
-    fn get_grid_size(&self) -> Result<(usize, usize), Error> {
-        return Err(Error::Unsupported);
-    }
-
     fn into_index(&self, event: Event) -> Result<Option<u16>, Error> {
          return match event {
             // filter "note down" events, for notes higher than C2 (36), and with strictly positive velocity
@@ -32,10 +28,6 @@ impl EventTransformer for DefaultEventTransformer {
     }
 
     fn into_color_palette_index(&self, _event: Event) -> Result<Option<u16>, Error> {
-        return Err(Error::Unsupported);
-    }
-
-    fn into_coordinates(&self, _event: Event) -> Result<Option<(u16, u16)>, Error> {
         return Err(Error::Unsupported);
     }
